Solve Problem 9.
Given the lowest pythagorean triplet of 3, 4, 5, start there, and
double each element.

Then, try to search for the next triplet that's between the current
and the doubled triplet, by slowly incrementing c, then b, then a.

If the sum of a b and c exceed the sumConstraint, then stop searching,
none could be found.

The product of the pythagorean triplet, whose terms add up to 1000,
is 31875000.