Work more efficiently with my enumerable ranges.
It turns out there's more I can do in a single enumeration
from 0 to 4 - I can add the rows across as well as the rows
down in the same enumeration.

For the rows down I still need to run an additional
enumeration because each cell down is from a different row
across; that's just how nested arrays work, at least as far
as I know.